+++ src/Stratis.Bitcoin.Features.PoA.Tests/VotingManagerTests.cs
@@ -30,7 +30,7 @@ namespace Stratis.Bitcoin.Features.PoA.Tests
         [Fact]
         public void CanScheduleAndRemoveVotes()
         {
-            this.federationManager.SetPrivatePropertyValue(typeof(FederationManagerBase), nameof(this.federationManager.IsFederationMember), true);
+            this.federationManager.SetPrivatePropertyValue(typeof(FederationManager), nameof(this.federationManager.IsFederationMember), true);
 
             this.votingManager.ScheduleVote(new VotingData());
