+++ src/Stratis.Features.FederatedPeg/Controllers/FederationGatewayClient.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Stratis.Bitcoin.Controllers;
 using Stratis.Bitcoin.Utilities;
-using Stratis.Features.Collateral.CounterChain;
 using Stratis.Features.FederatedPeg.Models;
+using Stratis.Features.PoA.Collateral.CounterChain;
 
 namespace Stratis.Features.FederatedPeg.Controllers
 {
